UI/GUI/requirements.txt
fastapi
uvicorn
requests
python-multipart
python-jose